Windows-specific Wordnet-2 dictionary. Better support for job editing. Support for transfering jobs scross organizations. Issue detections re. job supervisors. Minor fixes.
git-svn-id: https://wush.net/svn/mindalliance@1722 baad322d-9929-0410-88f0-f92e8ff3e1bd

+++ channels/trunk/src/main/java/com/mindalliance/channels/analysis/IssueScanner.java
@@ -90,9 +90,10 @@ public class IssueScanner implements Scanner {
      */
     public void terminate() {
         LOG.debug( "Terminating issue scans" );
-        for ( Daemon daemon : daemons.values() ) {
-            daemon.terminate();
-        }
+        if ( daemons != null )
+            for ( Daemon daemon : daemons.values() ) {
+                daemon.terminate();
+            }
     }
 
     /**
@@ -201,7 +202,7 @@ public class IssueScanner implements Scanner {
                 if ( !active ) return;
                 analyst.countTestFailures( getPlan(), Issue.ROBUSTNESS );
                 long endTime = System.currentTimeMillis();
-                LOG.info( "Issue sweep completed on " + getPlan() + " in " + (endTime - startTime) + " msecs" );
+                LOG.info( "Issue sweep completed on " + getPlan() + " in " + ( endTime - startTime ) + " msecs" );
             } catch ( Throwable e ) {
                 e.printStackTrace();
                 LOG.debug( "Deamon failed", e );
